Add common_getPos, split core/functions.hpp into multiple files
