+++ Missionframework/modules/01_core/fnc/common/fn_common_getPos.sqf
@@ -1,0 +1,41 @@
+#include "..\ui\defines.hpp"
+/*
+    KPLIB_fnc_common_getPos
+
+    File: fn_common_getPos.sqf
+    Author: KP Liberation Dev Team - https://github.com/KillahPotatoes
+    Date: 2018-09-12
+    Last Update: 2018-09-15
+    License: GNU General Public License v3.0 - https://www.gnu.org/licenses/gpl-3.0.html
+
+    Description:
+    Returns position of given marker, object or array.
+    In case array is given it is returned directly. For objects pos ATL is returned.
+
+    Parameter(s):
+        0: STRING/OBJECT/ARRAY - Position
+
+    Returns:
+    ARRAY - Position
+*/
+params [["_item", nil, ["",objNull,[],0], 3]];
+
+// Get position depending of type of respawn item
+switch (typeName _item) do {
+    case "STRING": {
+        getMarkerPos _item;
+    };
+
+    case "OBJECT": {
+        getPosATL _item;
+    };
+
+    // If someone forgots to wrap position array into an array.
+    case "SCALAR": {
+        +_this;
+    };
+
+    case "ARRAY": {
+       +_item;
+    };
+};
