+++ Missionframework/modules/01_core/fnc/common/functions.hpp
@@ -1,0 +1,28 @@
+/*
+    KP LIBERATION CORE FUNCTIONS
+
+    File: functions.hpp
+    Author: KP Liberation Dev Team - https://github.com/KillahPotatoes
+    Date: 2018-09-15
+    Last Update: 2018-09-15
+    License: GNU General Public License v3.0 - https://www.gnu.org/licenses/gpl-3.0.html
+
+    Description:
+    Defines for all common utility functions
+*/
+
+class common {
+    file = "modules\01_core\fnc\common";
+
+    // Clears vehicle cargo
+    class common_clearVehicleCargo;
+
+    // Generate positions in circle
+    class common_getCirclePositions {};
+
+    // getPos wrapper for ATL positions
+    class common_getPos {};
+
+    // Spawning of a vehicle at given position with given direction
+    class common_spawnVehicle {};
+};
